requirements.txt
asgiref==3.9.2
crispy-bootstrap5==2025.6
Django==5.2.6
django-crispy-forms==2.4
django-multiselectfield==1.0.1
django-quill-editor==0.1.42
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
